requirements.txt
langchain==0.3.9
langchain-google-genai==2.0.10
langchain-community==0.3.9
langchain-core==0.3.39
langchain-experimental==0.3.3
langchain-neo4j==0.1.1
neo4j==5.27.0
python-dotenv==1.0.1
json-repair
pypdf==4.3.1